requirements.txt
kafka-python-ng
pyspark
six
pandas